code cleanup/admin views in dashboard/admin auth products backend

+++ server/controllers/productController.js
@@ -3,11 +3,10 @@ const prisma = require("../prisma/PrismaClient")
 
 
 exports.getProducts = async (req, res) => {
-    try {     
-        console.log("Fetching Products")  
-        const products = await prisma.product.findMany({ include: { images: true, user:{select:{id:true,firstName:true, lastName:true, email:true}} }, orderBy:{title:'asc'} })
+    try {
+        console.log("Fetching Products")
+        const products = await prisma.product.findMany({ include: { images: true, user: { select: { id: true, firstName: true, lastName: true, email: true } } }, orderBy: { title: 'asc' } })
         prisma.$disconnect()
-        console.log(products)
         res.status(200).json(products)
     } catch (err) {
         console.log(err)
@@ -18,7 +17,7 @@ exports.getProducts = async (req, res) => {
 exports.getProductById = async (req, res) => {
     try {
         const { id } = req.params
-        const product = await prisma.product.findUnique({ where: { id: id }, include: { images: true,  user:{select:{firstName:true, lastName:true, email:true}} },  })
+        const product = await prisma.product.findUnique({ where: { id: id }, include: { images: true, user: { select: { firstName: true, lastName: true, email: true } } }, })
         res.status(200).json(product)
     } catch (err) {
         console.log(err)
@@ -26,15 +25,16 @@ exports.getProductById = async (req, res) => {
     }
 }
 
-exports.addProduct = async(req, res)=>{
-    try{
-         console.log("Adding New Product")
-         const product = {title:req.body.title, description:req.body.description, quantity:Number(req.body.quantity),price:Number(req.body.price), userId:req.body.userId}
+exports.addProduct = async (req, res) => {
+    try {
+        console.log("Adding New Product")
+        const product = { title: req.body.title, description: req.body.description, quantity: Number(req.body.quantity), price: Number(req.body.price), userId: req.body.userId }
         await prisma.product.create({
-            data:{...product,               
-                images:{
-                    create:{
-                        url:req.body.imageUrl
+            data: {
+                ...product,
+                images: {
+                    create: {
+                        url: req.body.imageUrl
                     }
                 }
             }
@@ -42,39 +42,52 @@ exports.addProduct = async(req, res)=>{
         prisma.$disconnect()
         console.log("Successfully Added Product")
         res.status(200).json("Successfully Added Product")
-    }catch(err){
+    } catch (err) {
         console.log(err)
         res.status(500).json(err)
     }
 }
 
-exports.editProduct = async(req, res)=>{
-    console.log("Editing Product")
-  
-    
-    try{
-         console.log("Editing Product")
-         const product = {title:req.body.title, description:req.body.description, quantity:Number(req.body.quantity),price:Number(req.body.price)}
-        await prisma.product.update({
-            where:{id:req.body.id},
-            data:{...product}
-        })
-        prisma.$disconnect()
-        console.log("Successfully Updated Product")
-        res.status(200).json("Successfully Updated Product")
-    }catch(err){
+exports.editProduct = async (req, res) => {
+    const user = req.user;
+   
+    try {
+        console.log("Editing Product")        
+        const product = await prisma.product.findUnique({ where: { id: req.body.id } })
+        const updatedProduct = { 
+            title: req.body.title, 
+            description: req.body.description, 
+            quantity: Number(req.body.quantity), 
+            price: Number(req.body.price), 
+            updatedAt:  new Date().toJSON()
+        }
+        
+        if (user.id === product.userId) {
+            await prisma.product.update({
+                where: { id: req.body.id },
+                data: { ...updatedProduct }
+            })
+            prisma.$disconnect()
+            console.log("Successfully Updated Product")
+            return res.status(200).json("Successfully Updated Product")
+        }
+        return res.status(403).json("You are not authorized to edit this product")
+    } catch (err) {
         console.log(err)
         res.status(500).json(err)
     }
 }
 
-exports.deleteProduct = async(req, res)=>{
-    try{
+exports.deleteProduct = async (req, res) => {
+    try {
         console.log("Deleting Product")
-        await prisma.productImage.deleteMany({where:{productId:req.params.id}})
-        await prisma.product.delete({where:{id:req.params.id}})
-        res.status(200).json("Successfully Delete Product")
-    }catch(err){
+        const product = await prisma.product.findUnique({where:{id:req.params.id}})
+        if(req.user.id === product.userId || req.user.role === 'ADMIN'){
+            await prisma.product.delete({ where: { id: req.params.id } })
+            return res.status(200).json("Successfully Delete Product")
+        }
+        return res.status(403).json("Unauthorized")      
+    } catch (err) {
         console.log(err)
         res.status(500).json(err)
     }
